Detection needs to be where it was expected
As a safety check we check if the marker were detected where it was
expected to be (within a certain threshold) to be considered a true match.
If it were detected on another place it is discarded from calibration.

Detection needs to be where it was expected

As a safety check we check if the marker were detected where it was
expected to be (within a certain threshold) to be considered a true match.
If it were detected on another place it is discarded from calibration.