Fix serious bug in psCHat2Rob
psCHat2Rob was previously calculated in the coordinate system of
psCHat2Rob (I think?!) but now it is reoriented to be in the robot
coordinate system. Unsure if it is in the direction from the CHat to the
robot or from the robot base to the CHat. But the coordinate system is
oriented according to the base at least.