Add separate function call for moving camera
Added separate function call for moving camera instead of switch,
this is because when moving the camera we may don't have another arm (the
marker can be located on a sticker on the wall) and therefore we cant get
any tooldata. Instead just return the position of the marker relative to
the robot base coordinate system.